requirements.txt
discord.py==2.4.0
pymongo==4.8.0
pytz==2024.1
